Move nonemin and nonemax into _reduction.py
