+++ python/src/scipp/plot/tools.py
@@ -203,18 +203,3 @@ def nonemin(a, b):
     if b is None:
         return a
     return min(a, b)
-
-
-def nonemax(a, b):
-    """
-    Return max(a, b) if neither arguments are None.
-    Otherwise return the argument that is not None.
-    :raises ValueError: If both a and b are None.
-    """
-    if a is None:
-        if b is None:
-            raise ValueError("One argument must not be None")
-        return b
-    if b is None:
-        return a
-    return max(a, b)
